# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sabah Park Api Documentation",
        "contact": {
            "email": "prastiyo.beka12@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/dashboard/summary": {
            "get": {
                "tags": [
                    "DASHBOARD"
                ],
                "summary": "Get Data Member",
                "operationId": "App\\Http\\Controllers\\Admin\\DashboardController::summary",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/member/logout": {
            "post": {
                "tags": [
                    "Member Credentials"
                ],
                "summary": "Logout Member",
                "operationId": "App\\Http\\Controllers\\MemberController::logout",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/member/profile": {
            "get": {
                "tags": [
                    "Member Credentials"
                ],
                "summary": "Get Profile Member",
                "operationId": "App\\Http\\Controllers\\MemberController::profile",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/member/change-password": {
            "post": {
                "tags": [
                    "Member Credentials"
                ],
                "summary": "Change Password Member",
                "operationId": "App\\Http\\Controllers\\MemberController::change_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/member/edit-profile": {
            "post": {
                "tags": [
                    "Member Credentials"
                ],
                "summary": "Edit Profile Member",
                "operationId": "App\\Http\\Controllers\\MemberController::edit_profile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "postcode": {
                                        "type": "string"
                                    },
                                    "nationality": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "1 = Pria, 2 = Wanita",
                                        "type": "string"
                                    },
                                    "tgl_lahir": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/public/negara/get": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Get Negara",
                "operationId": "App\\Http\\Controllers\\NegaraController::get",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/admin/profile": {
            "get": {
                "tags": [
                    "Admin Credentials"
                ],
                "summary": "Get Profile Admin",
                "operationId": "App\\Http\\Controllers\\UserController::profile",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/admin/change-password": {
            "post": {
                "tags": [
                    "Admin Credentials"
                ],
                "summary": "Change Password Admin",
                "operationId": "App\\Http\\Controllers\\UserController::change_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/admin/edit-profile": {
            "post": {
                "tags": [
                    "Admin Credentials"
                ],
                "summary": "Edit Profile Admin",
                "operationId": "App\\Http\\Controllers\\UserController::edit_profile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "fullname": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "postcode": {
                                        "type": "string"
                                    },
                                    "nationality": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "1 = Pria, 2 = Wanita",
                                        "type": "string"
                                    },
                                    "tgl_lahir": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}